增加springboot3.1.1-mybatisplus-redis-java17

+++ springcloud3.1.1-mybatisplus-redis-java17/$module$/src/main/java/$package$/$model$/ctrl/$className$Controller.java
@@ -57,7 +57,7 @@ public class $className$Controller {
         }
         /***}***/
 
-        int count = $classNameLower$Service.lambdaQuery()
+        long count = $classNameLower$Service.lambdaQuery()
         /***
          for(field in fields){
          ***/
